==> 02_process_WS24.ipynb <==
from CGCNN_MT.datamodule.prepare_data import prepare_data
from CGCNN_MT.datamodule.prepare_data import make_prepared_data, get_logger
from pathlib import Path
from tqdm import tqdm
import multiprocessing as mp
import shutil
from functools import partial
import pandas as pd
import json
import warnings
from sklearn.model_selection import train_test_split
import numpy as np
import shutil
warnings.filterwarnings("ignore", category=UserWarning, module="pymatgen.io.cif")
warnings.filterwarnings("ignore", category=UserWarning, module="moftransformer.utils.prepare_data")
warnings.filterwarnings("ignore", category=UserWarning, module="ase.io.cif")

src_csv = "./raw_data/WS24v2/features/features_and_labels.csv"
src_cif_dir1 = Path("./raw_data/WS24v2/data_sets/WS14s/CIFs")
src_cif_dir2 = Path("./raw_data/WS24v2/data_sets/WS24s/CIFs")
tgt_root_dir = Path("./CGCNN_MT/data/WS24")

tgt_root_dir.mkdir(exist_ok=True, parents=True)

print(len(list(src_cif_dir1.glob('*.cif'))))
print(len(list(src_cif_dir2.glob('*.cif'))))

in_df = pd.read_csv(src_csv)
in_df.head(2)

## Move cif files

mofs_withot_file = []
tgt_cif_dir = tgt_root_dir/"cifs"
# if tgt_cif_dir.exists():
#     shutil.rmtree(tgt_cif_dir)
tgt_cif_dir.mkdir(exist_ok=True)
for i in range(len(in_df)):
    n = in_df.iloc[i]["MOF_name"]
    dataset = in_df.iloc[i]["data_set"]
    if dataset == "WS14s":
        cif_file = src_cif_dir1/f"{n}.cif"
    elif dataset == "WS24s":
        cif_file = src_cif_dir2/f"{n}.cif"
    if not cif_file.exists():
        print(f"{cif_file} does not exist.")
        mofs_withot_file.append(n)
        continue
    shutil.copy(cif_file, tgt_cif_dir)

if len(mofs_withot_file) > 0:
    print(f"There are {len(mofs_withot_file)} MOFs without CIF files.")
    with open(tgt_root_dir/"mofs_withot_cif.txt", "w") as f:
        f.write("\n".join(mofs_withot_file))

### Reneragete zeo++ and RACs features for WS24 datasets.

from CGCNN_MT.datamodule import clean_cif
import subprocess
import os

## clean cifs
for task in  ["WS24"]:
    task_dir =tgt_root_dir/task
    cif_dir = task_dir/"cifs"
    output_dir = task_dir/"clean_cifs"
    clean_cif.main(cif_dir, output_dir, log_file=task_dir/"clean_cif.log", santize=True, n_cpus=1)

## generate features
for task in  ["WS24"]:
    task_dir =tgt_root_dir/task
    clean_cif_dir = task_dir/"clean_cifs"
    output_dir = task_dir/"features"
    work_dir = Path("./ML/featuring")
    process = subprocess.Popen(
        f"python {work_dir/'feature_generation.py'} --cif_dir {clean_cif_dir} --prob_radius 1.86",
        stdout=subprocess.PIPE,
        stderr=subprocess.PIPE,
        shell=True,
        env=os.environ.copy(),
        cwd=str(work_dir)
    )

## Prepare WS24 dataset for CGCNN_MT

import pickle
import os
root_dir = Path("./CGCNN_MT/data/WS24")
cif_dir = root_dir/"clean_cifs"
logger = get_logger(filename=str(root_dir/f"prepare_data_{root_dir.name}.log"))

# Generate prepared data for WS24 dataset
failed_records = []
for cif_file in tqdm(list(cif_dir.glob("*.cif"))):
    g_file_name = cif_file.stem + ".graphdata"
    if (cif_dir/g_file_name).exists():
        with open(cif_dir/g_file_name, 'rb') as f:
            cif_id, atom_num, nbr_idx, nbr_dist, *_, cell_params = pickle.load(f)
        if nbr_idx.shape[0] / atom_num.shape[0] != 10.0:
            print(f"Number of neighbors is not 10 for {g_file_name}")
            os.remove(str(cif_dir/g_file_name))
            flag = make_prepared_data(cif_file, cif_dir, logger=logger, max_num_nbr=10, radius=8)
            if not flag:
                # print(f"Failed to generate graph data for {cif_file}")
                failed_records.append(cif_file.stem)
    else:
        flag = make_prepared_data(cif_file, cif_dir, logger=logger, max_num_nbr=10, radius=8)
        if not flag:
            # print(f"Failed to generate graph data for {cif_file}")
            failed_records.append(cif_file.stem)
            continue
print("Failed records:", len(failed_records))

if len(failed_records) > 0:
    with open(tgt_root_dir/"mofs_prepare_failed.txt", "w") as f:
        f.write("\n".join(failed_records))

print(len(list(cif_dir.glob('*.cif'))))
print(len(list(cif_dir.glob('*.graphdata'))))

## Split the WS24 dataset into training, validation, and test sets

failed_records = []
in_df = pd.read_csv(src_csv)
in_df.insert(3, "water4_label", in_df["water_label"].astype(int))
in_df["water_label"] = in_df["water4_label"].apply(lambda x: 1 if x > 2 else 0)
in_df = in_df[~in_df['MOF_name'].isin(failed_records)]
in_df.rename(columns={"MOF_name": "MofName"}, inplace=True)
print(in_df.shape)
in_df.head(2)

def split_train_test(df, test_size=0.1, group_column="", random_state=0):
    """Split the dataset into train and test sets."""
    if group_column:
        train_index, test_index = train_test_split(df.index, test_size=test_size, stratify=df[group_column], random_state=random_state)
    else:
        train_index, test_index = train_test_split(df.index, test_size=test_size, random_state=random_state)
    print(f'Train test split successfully: train/test = {len(train_index)}/{len(test_index)}')
    return train_index, test_index


for lb in in_df["water4_label"].unique():
    print(lb, len(in_df[in_df["water4_label"] == lb]))

random_seed = 42

label_cols = ['water_label', 'acid_label', 'base_label', 'boiling_label']

train_df, test_df = train_test_split(in_df, test_size=0.2, stratify=in_df[label_cols], random_state=random_seed)

train_df, val_df = train_test_split(train_df, test_size=0.2, stratify=train_df[label_cols], random_state=random_seed)

train_df.insert(2, "Partition", "train")
val_df.insert(2, "Partition", "val")
test_df.insert(2, "Partition", "test")

total_df = pd.concat([train_df, val_df, test_df])
ratio_dict = {}
for col in label_cols + ["water4_label"]:
    for lb in total_df[col].unique():
        split_ratios = []
        for split in ["train", "val", "test"]:
            sub_df = total_df.loc[(total_df[col] == lb) & (total_df["Partition"] == split)]
            sub_ratio = len(sub_df) / len(total_df.loc[(total_df[col] == lb)])
            split_ratios.append([sub_ratio, len(sub_df)])
        ratio_dict[f"{col}_{lb}(train/val/test)"] = ' : '.join([f"{c}({r:.2f})" for r, c in split_ratios])
for k, v in ratio_dict.items():
    print(f"{k}: {v}")


total_df.head()

total_df.to_csv(tgt_root_dir/'id_prop_feat.csv', index=False)

### Reneragete zeo++ and RACs features for WS24 datasets.

df_new = pd.read_csv('./CGCNN_MT/data/WS24/RAC_and_zeo_features.csv')
df = pd.read_csv("./CGCNN_MT/data/WS24/id_prop_feat.csv")
# df.insert(3, "water4_label", df["water_label"])
# df["water_label"] = df["water4_label"].apply(lambda x: 0 if x<=2 else 1)
print(df_new.shape, df.shape)
df_new.drop(columns=['cif_file'], inplace=True)
df_new.rename(columns={'name': 'MofName'}, inplace=True)
df.rename(columns={'MOF_name': 'MofName', "split": "Partition"}, inplace=True)

id_prop_cols = ["MofName", "Partition", "water_label", "water4_label", "acid_label", "base_label", "boiling_label"]
df_new = df[id_prop_cols].merge(df_new, on="MofName", how="left")

df_new.to_csv("./CGCNN_MT/data/WS24/RAC_and_zeo_features_with_id_prop.csv", index=False)
