==> 05_ML_retrain.ipynb <==
%load_ext autoreload
%autoreload 2
from pathlib import Path
from ML.main import main
from ML.interface import ClassificationModel, RegressionModel

## WS24 training

work_dir = Path("./ML")
data_dir = work_dir / "data"


label_columns = ["water_label","water4_label", "acid_label", "base_label", "boiling_label"]
in_file_name="RAC_and_zeo_features_with_id_prop.csv"
for label_column in label_columns:
    task_name = "WS24"
    name_column="MofName"
    model_type = "classification"
    model_list = ["RF", "GP", "SVM", "LR"]
    feature_selector_list = ["RFE", "f1", "mutual_info"]
    search_max_evals = 100
    search_metric = "val_AUC"
    task_dir = data_dir/task_name
    saved_dir = work_dir/"models_test"/task_name/(in_file_name[:-4])/label_column
    main(
        data_dir=task_dir,
        in_file_name=in_file_name,
        label_column=label_column,
        name_column=name_column,
        model_type=model_type,
        model_list=model_list,
        feature_selector_list=feature_selector_list,
        search_max_evals=search_max_evals,
        search_metric=search_metric,
        saved_dir=saved_dir
    )

## TSD & SSD training

work_dir = Path("./ML")
data_dir = work_dir / "data"

task_names = ["TSD", "SSD"]
model_list = ["RF", "GP", "SVM", "LR"]
in_file_name = "RAC_and_zeo_features_with_id_prop.csv"
for task_name in task_names:
    label_column="Label"
    name_column="MofName"
    feature_selector_list = ["RFE", "f1", "mutual_info"]
    search_max_evals = 100
    if task_name == "TSD":
        model_type = "regression"
        search_metric = "val_R2"
    elif task_name in ["SSD", "WS24"]:
        model_type = "classification"
        search_metric = "val_AUC"
    task_dir = data_dir/task_name
    saved_dir = work_dir/"models_test"/task_name/(in_file_name[:-4])/label_column
    main(
        data_dir=task_dir,
        in_file_name=in_file_name,
        label_column=label_column,
        name_column=name_column,
        model_type=model_type,
        model_list=model_list,
        feature_selector_list=feature_selector_list,
        search_max_evals=search_max_evals,
        search_metric=search_metric,
        saved_dir=saved_dir
    )
