==> 06_ML_predicition_of_external_test_set.ipynb <==
%load_ext autoreload
%autoreload 2
import os
import sys
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from pathlib import Path
from sklearn import metrics
from ML.module import ClassificationModel, RegressionModel
sys.path.append("./ML")

external_data_root = Path("./CGCNN_MT/data/")
ts_external_csv = external_data_root/"TS_external_test/RAC_and_zeo_features_with_id_prop.csv"
ws_external_csv = external_data_root/"WS24v2_external_test/RAC_and_zeo_features_with_id_prop.csv"
log_dir = Path("./CGCNN_MT/evaluation")

df_ts = pd.read_csv(ts_external_csv)
df_ws = pd.read_csv(ws_external_csv)

print(df_ts.shape, df_ws.shape)

df_ts.head()

## Get best ML models

model_root_dir = "./ML/models"
model_root_dir = Path(model_root_dir)
cls_tasks = ["SSD", "WS24"]
reg_tasks = ["TSD"]
all_metrics = {}
for task_set, tp in zip([cls_tasks, reg_tasks], ["cls", "reg"]):
    dfs = []
    for task in task_set:
        model_dir = model_root_dir / task/ "RAC_and_zeo_features_with_id_prop"
        for subdir in model_dir.iterdir():
            if not subdir.is_dir():
                continue
            model_metrics_file = subdir / "model_metrics.csv"
            df_metrics = pd.read_csv(model_metrics_file, index_col=0)
            keep_cols = [col for col in df_metrics.columns if "tr_" not in col and "val_" not in col]
            df_metrics = df_metrics.loc[["mean"], keep_cols].copy()
            df_metrics.insert(0, "Task", "_".join([task, subdir.name]).replace("_label", "").replace("_Label", ""))
            dfs.append(df_metrics)
            if tp == "cls":
                df_metrics.sort_values(by=["te_AUC"], inplace=True, ascending=False)
            else:
                df_metrics.sort_values(by=["te_MAE"], inplace=True)
            df_metrics.columns = [col.replace("te_", "") for col in df_metrics.columns]
            all_metrics[task + "_" + subdir.name] = df_metrics.to_dict("records")[0]
df_ml_our = pd.DataFrame(all_metrics).T.copy()
df_ml_our

## Test on external test set

col2tasks = [
    {"ts_label": "TSD", "ss_label": "SSD"},
    {"water_label": "WS24_water", "water4_label": "WS24_water4", "acid_label": "WS24_acid", "base_label": "WS24_base", "boiling_label": "WS24_boiling"}
]
task2col = {"TSD": "Label", "SSD": "Label"}

task2col.update({t: t.split("_")[-1] + "_label" for t in ["WS24_water", "WS24_water4", "WS24_acid", "WS24_base", "WS24_boiling"]})
split = "external_test"
results = []
model_name = "Baseline"
evaluation_dir = log_dir/model_name
evaluation_dir.mkdir(exist_ok=True, parents=True)
for df, col2task in zip([df_ts, df_ws], col2tasks):
    for label_col, task in col2task.items():
        row = df_ml_our[df_ml_our["Task"] == task].iloc[0]
        model_file = f'total_model_{row["feature_selector"]}{row["feature_num"]}_none1_{row["model_name"]}_{row["random_state"]}.model'
        model_path = model_root_dir / task.split("_")[0] / "RAC_and_zeo_features_with_id_prop" / task2col[task] / model_file
        print(model_path)
        if not os.path.exists(model_path):
            print(f"{model_path} does not exist")
            continue
        if task == "TSD":
            model = RegressionModel(random_state=0)
            model.load_total_model(model_path)
            model.train(model.model, model.params)
            y_pred = model.predict(df.loc[:, "Di":]).squeeze()
            y_true = df[label_col].values
            r2, rmse, mae, pearson_corr, spearman_corr = model.cal_metrics(y_true, y_pred)
            results.append({"Task": task, "Model": model_name, "R2": r2, "MAE": mae})
            df_pred = pd.DataFrame([df["MofName"].values, y_true, y_pred, np.abs(y_true - y_pred)], 
                         index=["MofName", "GroundTruth", "Predicted", "Error"]).T
            df_pred.to_csv(evaluation_dir/f"{split}_results_{task}.csv", index=False)
        else:
            model = ClassificationModel(random_state=0, n_class=2)
            model.load_total_model(model_path)
            model.train(model.model, model.params)
            y_pred = model.predict(df.loc[:, "Di":], return_prob=False).squeeze()
            y_prob = model.predict(df.loc[:, "Di":], return_prob=True)
            y_true = df[label_col].values
            if task == "WS24_water4":
                y_true = y_true - 1
            acc, bacc, mcc, auc, f1, ce = model.cal_metrics(y_true, y_pred, y_prob, model.n_class)
            results.append({"Task": task, "Model": model_name, "ACC": acc, "BACC": bacc, "AUROC": auc})
            if y_prob.shape[1] == 2:
                y_prob = y_prob[:, 1]
            df_pred = pd.DataFrame([df["MofName"].values, y_true, y_pred, y_prob.tolist()], 
                         index=["MofName", "GroundTruth", "Predicted", "Prob"]).T
            df_pred.to_csv(evaluation_dir/f"{split}_results_{task}.csv", index=False)
df_ml_res = pd.DataFrame(results)
df_ml_res

df_ml_res["Model"] = "Baseline"
df_ml_res.set_index(keys=['Task', 'Model'], inplace=True)
df_ml_res.to_excel(log_dir/"ML_results_external_test.xlsx")

## Get Model hyperparameters

col2tasks = [
    {"ts_label": "TSD", "ss_label": "SSD"},
    {"water_label": "WS24_water", "water4_label": "WS24_water4", "acid_label": "WS24_acid", "base_label": "WS24_base", "boiling_label": "WS24_boiling"}
]
task2col = {"TSD": "Label", "SSD": "Label"}

task2col.update({t: t.split("_")[-1] + "_label" for t in ["WS24_water", "WS24_water4", "WS24_acid", "WS24_base", "WS24_boiling"]})
split = "external_test"
results = []
model_name = "Baseline"
evaluation_dir = log_dir/model_name
evaluation_dir.mkdir(exist_ok=True, parents=True)
for df, col2task in zip([df_ts, df_ws], col2tasks):
    for label_col, task in col2task.items():
        row = df_ml_our[df_ml_our["Task"] == task].iloc[0]
        model_file = f'total_model_{row["feature_selector"]}{row["feature_num"]}_none1_{row["model_name"]}_{row["random_state"]}.model'
        model_path = model_root_dir / task.split("_")[0] / "RAC_and_zeo_features_with_id_prop" / task2col[task] / model_file
        print("="*50)
        print("Model:", task + "_" + row["model_name"], "feature_selector:", row["feature_selector"], "feature_num:", row["feature_num"])
        if not os.path.exists(model_path):
            print(f"{model_path} does not exist")
            continue
        if task == "TSD":
            model = RegressionModel(random_state=0)
            model.load_total_model(model_path)
            print("model params:", model.params)
            
        else:
            model = ClassificationModel(random_state=0, n_class=2)
            model.load_total_model(model_path)
            print("model params:", model.params)
